refactor: update imports to use shared index.js
- Update Navigation Link block to import from shared index
- Update Navigation Submenu block to import from shared index
- Update Navigation block menu-inspector-controls to import from shared index
- Fix internal imports in shared controls.js to use relative paths
- Fix test file mock paths to match new structure
- All tests passing (53 tests in navigation-link, 86 total in navigation)
