+++ packages/block-library/src/navigation-link/shared/index.js
@@ -6,3 +6,4 @@
  */
 
 export { Controls } from './controls';
+export { updateAttributes } from './update-attributes';
